requirements.txt
attrs
black
biopython
hdbscan
jupyter
ndcctools
numpy
pandas
plotly
plotly-orca
psutil
scikit-learn
scikit-bio
star
tsne
umap-learn
bandage
samtools
bedtools
bowtie2